requirements.txt
mesop
google-genai
google-cloud-aiplatform
google-cloud-storage
google.auth
python-dotenv